+++ crates/ruma-events/src/ignored_user_list.rs
@@ -4,10 +4,10 @@ use ruma_events_macros::BasicEventContent;
 use ruma_identifiers::UserId;
 use serde::{Deserialize, Serialize};
 
-use crate::BasicEvent;
+use crate::GlobalAccountDataEvent;
 
 /// A list of users to ignore.
-pub type IgnoredUserListEvent = BasicEvent<IgnoredUserListEventContent>;
+pub type IgnoredUserListEvent = GlobalAccountDataEvent<IgnoredUserListEventContent>;
 
 /// The payload for `IgnoredUserListEvent`.
 #[derive(Clone, Debug, Deserialize, Serialize, BasicEventContent)]
@@ -34,11 +34,11 @@ mod tests {
     use serde_json::{from_value as from_json_value, json, to_value as to_json_value};
 
     use super::IgnoredUserListEventContent;
-    use crate::{AnyBasicEventContent, BasicEvent};
+    use crate::{AnyGlobalAccountDataEventContent, GlobalAccountDataEvent};
 
     #[test]
     fn serialization() {
-        let ignored_user_list_event = BasicEvent {
+        let ignored_user_list_event = GlobalAccountDataEvent {
             content: IgnoredUserListEventContent {
                 ignored_users: vec![user_id!("@carl:example.com")],
             },
@@ -68,12 +68,12 @@ mod tests {
         });
 
         assert_matches!(
-            from_json_value::<Raw<BasicEvent<AnyBasicEventContent>>>(json)
+            from_json_value::<Raw<GlobalAccountDataEvent<AnyGlobalAccountDataEventContent>>>(json)
                 .unwrap()
                 .deserialize()
                 .unwrap(),
-            BasicEvent {
-                content: AnyBasicEventContent::IgnoredUserList(IgnoredUserListEventContent {
+            GlobalAccountDataEvent {
+                content: AnyGlobalAccountDataEventContent::IgnoredUserList(IgnoredUserListEventContent {
                     ignored_users,
                 }),
             } if ignored_users == vec![user_id!("@carl:example.com")]